:recycle: api: move getAllWithSameTargetProfile to shared context
