+++ api/src/certification/shared/domain/models/ComplementaryCertificationBadge.js
