+++ api/tests/integration/domain/event/handle-complementary-certifications-scoring_test.js
@@ -4,10 +4,10 @@ import { CertificationScoringCompleted } from '../../../../lib/domain/events/Cer
 import { handleComplementaryCertificationsScoring } from '../../../../lib/domain/events/handle-complementary-certifications-scoring.js';
 import * as complementaryCertificationCourseResultRepository from '../../../../lib/infrastructure/repositories/complementary-certification-course-result-repository.js';
 import * as complementaryCertificationScoringCriteriaRepository from '../../../../lib/infrastructure/repositories/complementary-certification-scoring-criteria-repository.js';
-import * as complementaryCertificationBadgesRepository from '../../../../src/certification/complementary-certification/infrastructure/repositories/complementary-certification-badge-repository.js';
 import { AutoJuryCommentKeys } from '../../../../src/certification/shared/domain/models/JuryComment.js';
 import * as certificationAssessmentRepository from '../../../../src/certification/shared/infrastructure/repositories/certification-assessment-repository.js';
 import * as certificationCourseRepository from '../../../../src/certification/shared/infrastructure/repositories/certification-course-repository.js';
+import * as complementaryCertificationBadgesRepository from '../../../../src/certification/shared/infrastructure/repositories/complementary-certification-badge-repository.js';
 import { config as settings } from '../../../../src/shared/config.js';
 import { AnswerStatus } from '../../../../src/shared/domain/models/AnswerStatus.js';
 import * as assessmentResultRepository from '../../../../src/shared/infrastructure/repositories/assessment-result-repository.js';
